Tune .clang-format to match existing style
These settings seem to capture many of the qualities of the prevailing
style. The differences don't seem to be possible to resolve further.

All files in Imath, ImathTest, Half, and HalfTest have been reformatted.

Note that AlignConsecutiveDeclarations is set to false, because it
causes awkward formatting of template specializations that include
IMATH_CONSTEXPR14. The preprocessor symbol apparently makes it think
it's a declaration.

Formatting is disabled around some Matrix declarations, since
hand-formatting the initialization of a 4x4 matrix tends to look
better.

Note that this requires clang-format version 10.

Signed-off-by: Cary Phillips <cary@ilm.com>