Add .git-blame-ignore-revs to ignore reformatting
Use this file with git-blame to ignore commits that have reformatted
old code according to the new clang-format style:

  git blame --ignore-revs-file .git-blame-ignore-revs Imath/ImathVec.h

This helps preserve the old revision history, or at least makes it
more easily accessible.  Note that this requires git version 2.23 or
greater.

Also consider adding this to your git config:

  git config blame.ignoreRevsFile .git-blame-ignore-revs

Signed-off-by: Cary Phillips <cary@ilm.com>