Deprecate Math<T> in favor of std::
These were helpful to allow templated application of either double or
float in the olden days, but in C++11 and beyond, std:: versions of
these math functions do the same thing and should be preferred.

For back compatibility for applications, I'm keeping Math<T> but
defining them to just be the std:: versions. But the goal should be
to remove them eventually.

Signed-off-by: Larry Gritz <lg@larrygritz.com>