Imath::Vec -- omit exception-throwing methods from Cuda side
In Imath::Vec, remove IMATH_HOSTDEVICE from the few functions that
throw exceptions, since that's not supported in Cuda.

Interestingly the Matrix methods with exceptions were already
host-only (good job, Owen), so no changes were needed there.  When we
rig the rest of the library with `noexcept`, we'll be sure to double
check that there is no IMATH_HOSTDEVICE on the functions that still
throw exceptions.

Signed-off-by: Larry Gritz <lg@larrygritz.com>