noexcept all the things
This patch marks as noexcept absolutely everything that could be under
C++11 rules. Excluded are, obviously, anything that intentionally throws,
and those are very few after Owen's recent Imath overhaul and my changes
to eliminate the weird integer versions of normalize et al.

This patch also makes most of numeric_limits<half> constexpr, per
C++11 rules.  But a few are not constexpr yet because parts of half
still need to be fully constexpr-plumbed, and there are some tricky
parts (and maybe not 100% possible in C++11). I may revisit this in a
later patch.

Using `noexcept` is a promise that the caller to your function won't need
to handle exceptions. If your function calls something that turns out to
throw, it's on you to catch it, or else the application will crash with
an uncaught exception.

It's important that this get a thorough review. I really need a few
keen eyes to make sure that I haven't inadvertently marked anything
noexcept if it calls things (including outside Imath itself) that
might throw exceptions. (I mean, we can fix those bugs later, but I'd
prefer to find them now rather than in the heat of production.)

Signed-off-by: Larry Gritz <lg@larrygritz.com>