Extend interop to Vec2, Vec4, Matrix{22,33,44}
Also change the ones that require subscripting to non-constexpr,
non-noexcept, to accommodate interop with foreign types whose
operator[] is not constexpr or noexcept.  I have mixed feelings about
this; most sane implementations *ought* to be constexpr and noexcept,
but I can imagine that many are not so declared even though they could
be. I have left the .xyzw direct accessors constexpr and noexcept,
because simple assignment of these member variables ought to be safe.

Removed unused isSameType templates (if needed, we should use
`std::is_same<>`).

Signed-off-by: Larry Gritz <lg@larrygritz.com>