Interoperability constructors (#91)
Use enable_if to give `Vec3<T>` constructor and
assignment from unknown types that "look like" 3-vectors -- they have
a `operator[]` that returns a T, and their size is at least
3*sizeof(T).

The main idea is that if an app has rolled its own 3-vector class, it
can seamlessly construct a Vec3 from it, assign a Vec3 to it, or pass
their type as a parameter that expects a Vec3. And if the app's custom
3-vector class employs an equivalent idiom, all those operations will
work in the reverse direction.

It also works for std::vector, std::array, and even "initializer lists",
so all of this would work:

    // Your app's special snowflake custom 3-vector type.
    class YourCustomVec3 { ... };

    // My library has an API call where I use Imath::Vec3 as a parameter
    // passing convention, because I don't know about your vector.
    void myfunc (const Vec3<float>& v) { ... }

    // All sorts of things people may think of as "a vector"
    YourCustomVec3 myvec(1, 2, 3);
    std::array<float,3> arr { 1, 2, 3 };
    std::vector<float> stdvec { 1, 2, 3 };

    myfunc(yourvec);            // ok
    myfunc(arr);                // yep
    myfunc(stdvec);             // hunky-dory
    myfunc({ 1.0, 2.0, 3.0 });  // yeah, even that

This is only prototyped for Vec3 for now, but if you all like this
idea, then I will complete the work to do this for the other vector
and matrix classes.


Different approach:

* New file ImathTypeTraits.h now collects all the type traits (including
  a couple things we had previously put in ImathPlatform.h).

* New traits: has_subscript, has_xy, has_xyz, has_xyzw.

* New test file: testInterop.{h,cpp}

* Plain C arrays[3] work now.

* Classes that have .x, .y, .z member variables (but no subscripting)
  work now.

* Classes that have BOTH subscripting and named members work!

* It's easy to correct false positives and false negatives.

I still have only prototyped for Vec3, if people like the design, I will
extend to the other relevant types.

Signed-off-by: Larry Gritz <lg@larrygritz.com>

* Extend interop to Vec2, Vec4, Matrix{22,33,44}

Also change the ones that require subscripting to non-constexpr,
non-noexcept, to accommodate interop with foreign types whose
operator[] is not constexpr or noexcept.  I have mixed feelings about
this; most sane implementations *ought* to be constexpr and noexcept,
but I can imagine that many are not so declared even though they could
be. I have left the .xyzw direct accessors constexpr and noexcept,
because simple assignment of these member variables ought to be safe.

Removed unused isSameType templates (if needed, we should use
`std::is_same<>`).

Signed-off-by: Larry Gritz <lg@larrygritz.com>

* Some constexpr14 can be constexpr

Signed-off-by: Larry Gritz <lg@larrygritz.com>

* Revision

* Disable interop if the symbol IMATH_FOREIGN_VECTOR_INTEROP is
  defined to be 0 prior to including any Imath headers.

* Also disable if we detect gcc 4.x, which seems to have a bug that
  prevents this idiom from compiling correctly.

* Corral the interop functionality into a documentation "group" and
  beef up the description.

Signed-off-by: Larry Gritz <lg@larrygritz.com>

* Make the new constructors explicit

Signed-off-by: Larry Gritz <lg@larrygritz.com>

* Fix comments

Signed-off-by: Larry Gritz <lg@larrygritz.com>