Fix problems with IMATH_VERSION
The "IMATH_VERSION_EXTRA" cannot be incorporated into IMATH_VERSION
because CMake versions need to in the form be
major.minor[.patch[.tweak]] and IMATH_VERSION gets folded all the way
into ImathConfigVersion.cmake, and thus messes with downstream
projects.

I have now renamed "IMATH_VERSION_EXTRA" to "IMATH_VERSION_SUFFIX" for
clarity, and instead of appending it to IMATH_VERSION, I just
incorporate it into IMATH_PACKAGE_NAME, which can be an arbitrary
string. That will print the right designation when it's used, but without
polluting the numeric version.

Signed-off-by: Larry Gritz <lg@larrygritz.com>