Use .x instead of operator[] for better SIMD auto-vectorization
This cleans up a few more places where the use of operator[] is not as
helpful to compilers as using .x[]. They look like they should
generate the same code, and indeed they do for ordinary scalar
code. But when used inside a loop, it's enough to confuse the compiler
and prevent it from auto-vectorizing, and it can also be enough to
prevent use of `#pragma omp simd` from vectorizing loops.

We had largely done this for vector and matrix classes already, this
just tightens up a few spots we missed in the matrix classes.

These are based on suggestions by Alex Wells of Intel.

Signed-off-by: Larry Gritz <lg@larrygritz.com>