Move numeric_limits<half> specializations into half.h (#241)
Previously, these specializations have been in the header halfLimit.h,
but without this header, the numeric_limits<> template provides
default specializations that return 0, so, for example,
std::numeric_limits<half>::min() would silently return 0 in code that
inadvertently omitted the #include <halfLimits.h>.

The numeric_limits specializations should be considered an inseparable
part of the class definition.

This change moves the specializations directly into half.h and
deprecates the halfLimits.h header altogether. For backwards
compatibility, it includes half.h with a warning.

An alternative would be to keep halfLimits.h and include it in half.h,
but then each header would include the other circularly, which is
awkward.

Signed-off-by: Cary Phillips <cary@ilm.com>