Do not warn if half.h has already being included
The issue with the warning of halfLimits.h being deprecated is that it
is very tempting to simply remove the include of halfLimits.h in favor
of simply half.h. However this change does depends on the version of
Imath used, and halfLimits.h should be kept for compatibility with
older version. Simply do not warn if short.h is already included.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>