Conditionally include Matrix constructors duplicated by interop declarations
From the code coverage analysis, the ``Matrix44<>(const T a[4][4])``
constructors are overshadowed by the interoperability constructors
that match types with double-index operators. This change includes
them only if the interoperability constructors are disabled at compile
time.

It also adds a test to validate compilation with the interoperability
constructurs disabled.

Signed-off-by: Cary Phillips <cary@ilm.com>