Move test_m44.py to todo_m44.py until we have pybind11 bindings and can carefully compare the results with the boost python bindings.
Add a couple more operators to the box that we will definitely need for future
tests.

Add a bunch of operators to PyImathVec to get almost all of the tests passing.

Need to check on the argument order for closestVertex and project as the
boost python bindings move them around without a clear doc string indicating
it.

Signed-off-by: Lucas Miller <lmiller@imageworks.com>