Reoganize pybind11 cmake configuration (#434)
This reconfigures the new pybind11 implementation so it can be built
in parallel with the existing Boost.Python bindings, and arranges so
that the Boost.Python bindings can be deprecated and removed once with
new bindings are functional.

It's now possible to build the pybind11-based bindings without
requiring Boost.Python to be installed.

The CMake option ``PYTHON`` specifies whether to build the
Boost.Python-based bindings, same as always. The Boost.Python binding
source code is under src/python.

The new CMake optin ``PYBIND11`` specifies whether to build the
pybind11-based bindings. The pybind11 binding source code is under
src/pybind11. The bindings build into a module called "pybindimath"
and a library called "libPyBindImath.so". These are temporary names
until these bindings become official, at which time they'll be renamed
to "imath".

The new pybind11-based CMake configuration is simplfied, with all the
setup inline in src/pybind11/PyBindImath/CMakeLists.txt. It does not
rely on any of the sort of CMake functions or setups that the
Boost.Python bindings pull from src/python/config. That code is overly
complicated, and overly general since it's only used for two libraries
(PyImathand PyImathNumpy).

Signed-off-by: Cary Phillips <cary@ilm.com>