Fix undefined behavior in operator[]
If you take the address of a member, the pointer is only valid within
the size of that member, such that accesses past that (i.e. to get to y
or z of a vec3) are undefined behavior. by using this with a reinterpret
cast, that tells the compiler the pointer space contains all values, and
so avoids that. However, this may prevent other optimizations, so a
larger change is recommended.

Signed-off-by: Kimball Thurston <kdt3rd@gmail.com>