Add branchless access in an indexed way
When direct .x/.y/.z access is not possible, and subscript operators are
needed, this should enable the auto-vectorize optimization passes to
have a chance at making more efficient pipelines. This access method
change will also enable temporary stack vectors to be elided and live
entirely in registers and not need to be stored to the stack.

Signed-off-by: Kimball Thurston <kdt3rd@gmail.com>