Properly set PATH for CI tests (#497)
The "|| true" should have been removed in the earlier PR. It caused
the tests to succeed even though the PATH was incorrect. This sets the
PATH correctly.

Also, the Visual Studio build directory uses "link.command.1.tlog"
instead of "link.txt" for the link log, used to validate that there is
no dependency on PyImath.

Signed-off-by: Cary Phillips <cary@ilm.com>