Reset default install folder for PyImath to include/Imath
The default inadvertently changed to include/PyImath in the v3.2 release.

The significant change here to properly set the default for
`PYIMATH_OUTPUT_SUBDIR` and use it in `src/python/PyImath/CMakeLists.txt`.

The changes in the install manifests simply correct the mistaken
install locations. Note that the 'custom namespace' job specifies a
specific value of the subdir, to validate that the setting gets
properly overridden.

Signed-off-by: Cary Phillips <cary@ilm.com>